Remove pandas by using csv, datetime directly
